+++ src/cmd/mod.rs
@@ -1,6 +1,5 @@
 //! CLI commands.
 
-pub mod env;
 pub mod flash;
 pub mod gdb;
 pub mod heap;
@@ -10,6 +9,6 @@ pub mod reset;
 pub mod support;
 
 pub use self::{
-    env::run as env, flash::run as flash, gdb::run as gdb, heap::run as heap, log::run as log,
-    new::run as new, reset::run as reset, support::run as support,
+    flash::run as flash, gdb::run as gdb, heap::run as heap, log::run as log, new::run as new,
+    reset::run as reset, support::run as support,
 };