+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Possible log types:
 
 ### Unreleased
 
+- [added] Extra memory blocks can be added to generated linker scripts by
+  providing custom blocks in form of `[memory.foo]` in `Drone.toml`
 - [added] Generated linker scripts are now available at
   `target/<target>/layout.ld.<stage>` paths
 - [added] Added `drone print target` command